[neat] refactoring to reflect structure used in other providers
